requirements.txt
ipykernel
langchain
langchain-community
pypdf
langchain-google-genai
pinecone
fastapi
uvicorn[standard]
gunicorn uvicorn[standard]
python-multipart
from_root
jinja2
python-dotenv
-e .
